+++ src/main/java/com/telepathicgrunt/tact/Config.java
@@ -7,26 +7,26 @@ import net.minecraftforge.fml.event.config.ModConfigEvent;
 public class Config {
     private static final ForgeConfigSpec.Builder BUILDER = new ForgeConfigSpec.Builder();
 
-    private static final ForgeConfigSpec.BooleanValue ADJUST_STRUCTURE_NBT_FILES = BUILDER
+    private static final ForgeConfigSpec.BooleanValue APPLY_TAG_ADJUSTMENTS = BUILDER
             .comment("""
-                    Whether to replace Alex's Caves structure NBT files with ones that fit a bit better with terrain.
+                    Whether to activate an internal datapack to add Alex's Caves biomes and blocks to certain tags for better mod compat.
                     """)
-            .define("adjustStructureNbtFiles", true);
+            .define("applyTagAdjustments", true);
 
     private static final ForgeConfigSpec.BooleanValue REPLACEABLE_SMALL_PLANTS = BUILDER
             .comment("""
-                    Whether to make Bone Worms, Underweed, Tree Star, Fiddlehead, Curly Fern, and Pewen Pines be
+                    Whether to make Underweed, Tree Star, Fiddlehead, and Curly Fern be
                     replaceable by another block when right clicking them while holding a block in your hands.
                     """)
             .define("replaceableSmallPlants", true);
 
     static final ForgeConfigSpec SPEC = BUILDER.build();
 
-    public static boolean adjustStructureNbtFiles;
+    public static boolean applyTagAdjustments;
     public static boolean replaceableSmallPlants;
 
     static void onLoad(final ModConfigEvent event) {
-        adjustStructureNbtFiles = ADJUST_STRUCTURE_NBT_FILES.get();
+        applyTagAdjustments = APPLY_TAG_ADJUSTMENTS.get();
         replaceableSmallPlants = REPLACEABLE_SMALL_PLANTS.get();
     }
 }