+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-### **(V.1.0.0 Changes) (1.20.1 Minecraft)**
+### **(V.1.0.1 Changes) (1.20.1 Minecraft)**
 
-#### Major:
-Initial release
+#### Tags:
+Removed and adjusted some tags as Alex's Caves mod now has updated to include many tags.
+
+Moved this mod's own tag adjustments into an internal datapack. Tag adjustments now has a config to allow turning them all on or off.
+
+#### Structures:
+Removed internal NBT file datapack as Alex's Caves updated to include the files themselves.